refactor: single home for keep-alive timestamps
